Add real support for ES 2015 classes. In ES2015, class constructors are special in that calling them without the `new` keyword triggers an exception. This is also true when using `call` and `apply` on them. Unfortunately, that means newless's current technique for constructing objects won't work when you use the actual class syntax like so:
  var Point = newless(class Point {
      constructor(x, y) {
          this.x = x;
          this.y = y;
      }
  });

SO! This change adds support for this by actually using the `new` keyword when a class is involved. To do so, it uses the spread operator or, if the spread operator isn't supported, by creating a function that will call `new` with the proper arguments. That fallback is pretty slow and bad, though -- it means creating a throwaway function and calling it every time an object is constructed. Ugh. But I can't think of another workable approach.

To mitigate the performance problem, this patch also attempts to determine if it is dealing with a class or a normal function by calling `toString()` on the function it is working with. This works the vast majority of the time, but it's worth noting that it can fail:

  - If `Function.prototype.toString` has been overwritten.
  - If newless is called on the class's constructor rather than the class itself, e.g:
    `newless(SomeClass.prototype.constructor)` rather than `newless(SomeClass)`.